requirements.txt
discord.py==1.7.3
Django==5.0.4
requests==2.31.0
asyncio==3.4.3
djangorestframework==3.15.1
